requirements.txt
fastapi>=0.75.0
uvicorn>=0.17.6
sqlalchemy>=1.4.36
psycopg2-binary>=2.9.3
kafka-python>=2.0.2
pydantic>=1.8.2
pydantic-settings>=2.0.0